ImageReFL/requirements.txt
accelerate
clip
datasets
diffusers>=0.16.0
torch
torchvision
torchmetrics
transformers
numpy==1.26.4
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
hpsv2
pytorch-fid

black
isort
pre-commit

lpips

git+https://github.com/openai/CLIP.git
